PATCH /module/:module PATCH /module/:module/:pipe

+++ auth/auth.go
@@ -35,6 +35,6 @@ func Authorize(c *gin.Context) {
 	}
 
 	c.Set("uuid", claims["sub"].(string))
-	c.Set("level", claims["level"].(int))
+	c.Set("level", int(claims["level"].(float64)))
 	c.Next()
 }
